cmbs/python/requirements.txt
fastapi==0.85.0
uvicorn==0.19.0
pydantic==1.10.15
asyncpg==0.27.0
sentence-transformers==2.6.0
huggingface_hub
transformers==4.39.1
torch
python-dotenv
python-multipart
starlette
starlette-session
itsdangerous
google-generativeai
markdown
aioredis
fastapi-sessions
starsessions
starsessions[redis]
langchain
beautifulsoup4
pandas
lxml
spacy